Added variance_fold_level for plsicox(). Added 'return_data' parameter for all cv methods. Miscellaneous changes.

+++ R/HDcox_plsicox.R
@@ -444,6 +444,7 @@ plsicox <- function (X, Y,
 #' @param fast_mode Logical. If fast_mode = TRUE, for each run, only one fold is evaluated simultaneously. If fast_mode = FALSE, for each run, all linear predictors are computed for test observations. Once all have their linear predictors, the evaluation is perform across all the observations together (default: FALSE).
 #' @param MIN_EPV Minimum number of Events Per Variable you want reach for the final cox model. Used to restrict the number of variables can appear in cox model. If the minimum is not meet, the model is not computed.
 #' @param return_models Logical. Return all models computed in cross validation.
+#' @param returnData Logical. Return original and normalized X and Y matrices.
 #' @param tol Numeric. Tolerance for solving: solve(t(P) %*% W)
 #' @param PARALLEL Logical. Run the cross validation with multicore option. As many cores as your total cores - 1 will be used. It could lead to higher RAM consumption.
 #' @param verbose Logical. If verbose = TRUE, extra messages could be displayed (default: FALSE).
@@ -457,12 +458,12 @@ cv.plsicox <- function (X, Y,
                         n_run = 10, k_folds = 10,
                         x.center = TRUE, x.scale = FALSE,
                         y.center = FALSE, y.scale = FALSE,
-                        remove_near_zero_variance = T, remove_zero_variance = T, toKeep.zv = NULL,
+                        remove_near_zero_variance = T, remove_zero_variance = T, toKeep.zv = NULL, remove_variance_at_fold_level = F,
                         remove_non_significant_models = F, remove_non_significant = F, alpha = 0.05,
                         w_AIC = 0, w_c.index = 0, w_AUC = 1, times = NULL,
                         MIN_AUC_INCREASE = 0.05, MIN_AUC = 0.8, MIN_COMP_TO_CHECK = 3,
                         pred.attr = "mean", pred.method = "cenROC", fast_mode = F,
-                        MIN_EPV = 5, return_models = F, tol = 1e-15,
+                        MIN_EPV = 5, return_models = F, returnData = F, tol = 1e-15,
                         PARALLEL = F, verbose = F, seed = 123){
 
   t1 <- Sys.time()
@@ -489,13 +490,18 @@ cv.plsicox <- function (X, Y,
   max.ncomp <- check.maxPredictors(X, Y, MIN_EPV, max.ncomp)
 
   #### REQUIREMENTS
-  lst_dnz <- deleteZeroOrNearZeroVariance(X = X,
-                                          remove_near_zero_variance = remove_near_zero_variance,
-                                          remove_zero_variance = remove_zero_variance,
-                                          toKeep.zv = toKeep.zv,
-                                          freqCut = 95/5)
-  X <- lst_dnz$X
-  variablesDeleted <- lst_dnz$variablesDeleted
+  if(!remove_variance_at_fold_level){
+    lst_dnz <- deleteZeroOrNearZeroVariance(X = X,
+                                            remove_near_zero_variance = remove_near_zero_variance,
+                                            remove_zero_variance = remove_zero_variance,
+                                            toKeep.zv = toKeep.zv,
+                                            freqCut = 95/5)
+    X <- lst_dnz$X
+    variablesDeleted <- lst_dnz$variablesDeleted
+  }else{
+    variablesDeleted <- NULL
+  }
+
 
   #### #
   # CV #
@@ -513,24 +519,25 @@ cv.plsicox <- function (X, Y,
   #### ### ### ###
 
   total_models <- 1 * k_folds * n_run #with greatest component we have all of them
-
+  t1 <- Sys.time()
   comp_model_lst <- get_HDCOX_models2.0(method = pkg.env$plsicox,
                                        lst_X_train = lst_X_train, lst_Y_train = lst_Y_train,
                                        max.ncomp = max.ncomp, eta.list = NULL, EN.alpha.list = NULL, n_run = n_run, k_folds = k_folds,
                                        x.center = x.center, x.scale = x.scale, y.center = y.center, y.scale = y.scale,
-                                       remove_near_zero_variance = F, remove_zero_variance = F, toKeep.zv = NULL,
+                                       remove_near_zero_variance = ifelse(remove_variance_at_fold_level,T,F), remove_zero_variance = ifelse(remove_variance_at_fold_level,T,F), toKeep.zv = NULL,
                                        remove_non_significant = remove_non_significant,
                                        total_models = total_models, tol = tol, PARALLEL = PARALLEL, verbose = verbose)
-
+  t2 <- Sys.time()
+  t2-t1
   if(all(is.na(unlist(comp_model_lst)))){
-    message(paste0("Best model could NOT be obtained. All models computed present problems."))
+    message(paste0("Best model could NOT be obtained. All models computed present problems. Try to remove variance at fold level. If problem persists, try to delete manually some problematic variables."))
 
     t2 <- Sys.time()
     time <- difftime(t2,t1,units = "mins")
     if(return_models){
-      return(cv.plsicox_class(list(best_model_info = NULL, df_results_folds = NULL, df_results_runs = NULL, df_results_comps = NULL, lst_models = comp_model_lst, pred.method = pred.method, opt.comp = NULL, plot_AUC = NULL, plot_c_index = NULL, plot_AIC = NULL, class = pkg.env$cv.plsicox, time = time)))
+      return(cv.plsicox_class(list(best_model_info = NULL, df_results_folds = NULL, df_results_runs = NULL, df_results_comps = NULL, lst_models = comp_model_lst, pred.method = pred.method, opt.comp = NULL, plot_AUC = NULL, plot_c_index = NULL, plot_AIC = NULL, nzv = variablesDeleted, class = pkg.env$cv.plsicox, time = time)))
     }else{
-      return(cv.plsicox_class(list(best_model_info = NULL, df_results_folds = NULL, df_results_runs = NULL, df_results_comps = NULL, lst_models = NULL, pred.method = pred.method, opt.comp = NULL, plot_AUC = NULL, plot_c_index = NULL, plot_AIC = NULL, class = pkg.env$cv.plsicox, time = time)))
+      return(cv.plsicox_class(list(best_model_info = NULL, df_results_folds = NULL, df_results_runs = NULL, df_results_comps = NULL, lst_models = NULL, pred.method = pred.method, opt.comp = NULL, plot_AUC = NULL, plot_c_index = NULL, plot_AIC = NULL, nzv = variablesDeleted, class = pkg.env$cv.plsicox, time = time)))
     }
   }
